add suport for discord url scheme
this will allow the discord website to detect and hand over actions to
the app such as joining a server through a invite link.

these are the same arguments that are used to implement the feature on
windows

implemented as a separate hidden desktop file to keep the main file
without any extra arguments.

this feature request was also send to discord support, so may not be
necessary to do all of this in the future.